API Reference

Public

LinAlgTools.ref! — Method

ref!(A; show_steps=false)

Put A in row echelon form.

Returns the (row, column) indices of the pivots.

Example

julia> A = Float64[mod(i+j,3) for i in 1:4, j in 1:6]
4×6 Array{Float64,2}:
 2.0  0.0  1.0  2.0  0.0  1.0
 0.0  1.0  2.0  0.0  1.0  2.0
 1.0  2.0  0.0  1.0  2.0  0.0
 2.0  0.0  1.0  2.0  0.0  1.0

julia> ref!(A)
3-element Array{Tuple{Int64,Int64},1}:
 (1, 1)
 (2, 2)
 (3, 3)

julia> A
4×6 Array{Float64,2}:
 2.0  0.0   1.0  2.0  0.0   1.0
 0.0  1.0   2.0  0.0  1.0   2.0
 0.0  0.0  -4.5  0.0  0.0  -4.5
 0.0  0.0   0.0  0.0  0.0   0.0

LinAlgTools.row_add! — Method

row_add!(A, i, λ, i₀)

Add to row i of matrix A row i₀ multiplied by λ.

Returns the new row i.

Example

julia> A = [i//j for i in 1:3, j in 1:4]
3×4 Array{Rational{Int64},2}:
 1//1  1//2  1//3  1//4
 2//1  1//1  2//3  1//2
 3//1  3//2  1//1  3//4

julia> row_add!(A, 3, -3, 1)
4-element Array{Rational{Int64},1}:
 0//1
 0//1
 0//1
 0//1

julia> A
3×4 Array{Rational{Int64},2}:
 1//1  1//2  1//3  1//4
 2//1  1//1  2//3  1//2
 0//1  0//1  0//1  0//1

LinAlgTools.row_mul! — Method

row_mul!(A, i, λ)

Multiply row i of matrix A by a factor λ.

Returns the new row i.

Example

julia> A = reshape(1:12, 3, 4) |> collect
3×4 Array{Int64,2}:
 1  4  7  10
 2  5  8  11
 3  6  9  12

julia> row_mul!(A, 2, 100)
4-element Array{Int64,1}:
  200
  500
  800
 1100

julia> A
3×4 Array{Int64,2}:
   1    4    7    10
 200  500  800  1100
   3    6    9    12

LinAlgTools.row_swap! — Method

row_swap!(A, i, j)

Swap rows i and j in matrix A.

Returns the two rows in the new order.

Example

julia> A = reshape(1:12, 3, 4) |> collect
3×4 Array{Int64,2}:
 1  4  7  10
 2  5  8  11
 3  6  9  12

julia> row_swap!(A, 1, 3)
2×4 Array{Int64,2}:
 3  6  9  12
 1  4  7  10

julia> A
3×4 Array{Int64,2}:
 3  6  9  12
 2  5  8  11
 1  4  7  10

LinAlgTools.rref! — Method

rref!(A; show_steps)

Compute the reduced row echelon form of A.

Returns the (row, column) indices of the pivots.

Example

julia> A = Float64[mod(i+j,3) for i in 1:4, j in 1:6]
4×6 Array{Float64,2}:
 2.0  0.0  1.0  2.0  0.0  1.0
 0.0  1.0  2.0  0.0  1.0  2.0
 1.0  2.0  0.0  1.0  2.0  0.0
 2.0  0.0  1.0  2.0  0.0  1.0

julia> rref!(A)
3-element Array{Tuple{Int64,Int64},1}:
 (1, 1)
 (2, 2)
 (3, 3)

julia> A
4×6 Array{Float64,2}:
  1.0   0.0  0.0   1.0   0.0  0.0
  0.0   1.0  0.0   0.0   1.0  0.0
 -0.0  -0.0  1.0  -0.0  -0.0  1.0
  0.0   0.0  0.0   0.0   0.0  0.0

Private

LinAlgTools.log — Method

log(A, txt)

Show matrix A with title text txt.

LinAlgTools.ref_pass! — Method

ref_pass!(A)

Perform one pass of the row-echelon-form algorithm.

Returns the column index of the pivot for this pass, or nothing if A is all zeros.